+++ CHANGELOG.md
@@ -31,6 +31,9 @@
 
 ### Breaking changes
 
+* nostr: update `Tags::new` signature ([Yuki Kishimoto])
+* nostr: remove `WeakTag` ([Yuki Kishimoto])
+
 ### Changed
 
 * nostr: manually impl eq and cmp traits for `RelayUrl` ([Yuki Kishimoto])
@@ -49,7 +52,12 @@
 
 ### Added
 
+* nostr: add `nip21::extract_from_text` ([Yuki Kishimoto])
 * nostr: add `EventBuilder::allow_self_tagging` ([Yuki Kishimoto])
+* nostr: add `Tag::len` ([Yuki Kishimoto])
+* nostr: add `push`, `pop`, `insert`, `remove` and `extend` methods to `Tags` struct ([Yuki Kishimoto])
+* nostr: add `Tags::from_list`, `Tags::with_capacity` and `Tags::parse` constructors ([Yuki Kishimoto])
+* nostr: add `Tags::dedup` ([Yuki Kishimoto])
 * pool: event verification cache ([Yuki Kishimoto])
 * ffi: add Mac Catalyst support in Swift package ([Yuki Kishimoto])